app/requirements.txt
cassandra-driver==3.25.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
geomet==0.2.1.post1
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
six==1.16.0
thrift==0.15.0
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.7.0
gunicorn==20.1.0
